+++ simple_manga_downloader/SMD.py
@@ -213,10 +213,9 @@ def main_pipeline(links):
     if ARGS.check_only or ARGS.ignore_input:
         modules.toggle_check_only()
 
-    data_saver = getattr(ARGS, "data_saver", None)
-    if data_saver is None:
+    if ARGS.data_saver is None:
         modules.set_data_saver(CONFIG.data_saver)
-    elif data_saver == "true":
+    elif ARGS.data_saver == "true":
         modules.set_data_saver(True)
     else:
         modules.set_data_saver(False)